+++ dart/test/protocol/sentry_log_attribute_test.dart
@@ -11,8 +11,8 @@ void main() {
     });
   });
 
-  test('$SentryLogAttribute boolean to json', () {
-    final attribute = SentryLogAttribute.boolean(true);
+  test('$SentryLogAttribute bool to json', () {
+    final attribute = SentryLogAttribute.bool(true);
     final json = attribute.toJson();
     expect(json, {
       'value': true,
@@ -20,8 +20,8 @@ void main() {
     });
   });
 
-  test('$SentryLogAttribute integer to json', () {
-    final attribute = SentryLogAttribute.integer(1);
+  test('$SentryLogAttribute int to json', () {
+    final attribute = SentryLogAttribute.int(1);
     final json = attribute.toJson();
 
     expect(json, {