+++ dart/test/sentry_client_test.dart
@@ -23,6 +23,7 @@ import 'mocks/mock_client_report_recorder.dart';
 import 'mocks/mock_hub.dart';
 import 'mocks/mock_transport.dart';
 import 'test_utils.dart';
+import 'utils/url_details_test.dart';
 
 void main() {
   group('SentryClient captures message', () {
@@ -1699,6 +1700,201 @@ void main() {
       expect(envelopeEvent?.type, 'feedback');
       expect(envelopeEvent?.contexts.feedback?.toJson(), feedback.toJson());
       expect(envelopeEvent?.level, SentryLevel.info);
+    });
+  });
+
+  group('SentryClient captureLog', () {
+    late Fixture fixture;
+
+    setUp(() {
+      fixture = Fixture();
+    });
+
+    SentryLog givenLog() {
+      return SentryLog(
+        timestamp: DateTime.now(),
+        traceId: SentryId.newId(),
+        level: SentryLogLevel.info,
+        body: 'test',
+        attributes: {
+          'attribute': SentryLogAttribute.string('value'),
+        },
+      );
+    }
+
+    test('disabled by default', () async {
+      final client = fixture.getSut();
+      final log = givenLog();
+
+      await client.captureLog(log);
+
+      expect((fixture.transport).logs, isEmpty);
+    });
+
+    test('should capture logs as envelope', () async {
+      fixture.options.enableLogs = true;
+
+      final client = fixture.getSut();
+      final log = givenLog();
+      final logJson = log.toJson();
+
+      await client.captureLog(log);
+
+      final envelopePayloadJson = (fixture.transport).logs.first;
+
+      expect(envelopePayloadJson, isNotNull);
+      expect(envelopePayloadJson['items'].first['timestamp'],
+          logJson['timestamp']);
+      expect(
+          envelopePayloadJson['items'].first['trace_id'], logJson['trace_id']);
+      expect(envelopePayloadJson['items'].first['level'], logJson['level']);
+      expect(envelopePayloadJson['items'].first['body'], logJson['body']);
+      expect(
+          envelopePayloadJson['items'].first['attributes']['attribute']
+              ['value'],
+          'value');
+    });
+
+    test('should add additional info to attributes', () async {
+      fixture.options.enableLogs = true;
+      fixture.options.environment = 'test-environment';
+      fixture.options.release = 'test-release';
+
+      final log = givenLog();
+
+      final scope = Scope(fixture.options);
+      final span = MockSpan();
+      scope.span = span;
+
+      final client = fixture.getSut();
+      await client.captureLog(log, scope: scope);
+
+      final envelopePayloadJson = (fixture.transport).logs.first;
+      final attributesJson = envelopePayloadJson['items'].first['attributes'];
+
+      expect(
+        attributesJson['sentry.sdk.name']['value'],
+        fixture.options.sdk.name,
+      );
+      expect(
+        attributesJson['sentry.sdk.name']['type'],
+        'string',
+      );
+      expect(
+        attributesJson['sentry.sdk.version']['value'],
+        fixture.options.sdk.version,
+      );
+      expect(
+        attributesJson['sentry.sdk.version']['type'],
+        'string',
+      );
+      expect(
+        attributesJson['sentry.environment']['value'],
+        fixture.options.environment,
+      );
+      expect(
+        attributesJson['sentry.environment']['type'],
+        'string',
+      );
+      expect(
+        attributesJson['sentry.release']['value'],
+        fixture.options.release,
+      );
+      expect(
+        attributesJson['sentry.release']['type'],
+        'string',
+      );
+      expect(
+        attributesJson['sentry.trace.parent_span_id']['value'],
+        span.context.spanId.toString(),
+      );
+      expect(
+        attributesJson['sentry.trace.parent_span_id']['type'],
+        'string',
+      );
+    });
+
+    test('should set trace id if there is a scope span', () async {
+      fixture.options.enableLogs = true;
+
+      final client = fixture.getSut();
+      final log = givenLog();
+      final scope = Scope(fixture.options);
+      final span = MockSpan();
+      scope.span = span;
+
+      await client.captureLog(log, scope: scope);
+
+      final envelopePayloadJson = (fixture.transport).logs.first;
+      final logJson = envelopePayloadJson['items'].first;
+
+      expect(logJson['trace_id'], span.context.traceId.toString());
+    });
+
+    test(
+        'should set trace id from propagation context if there is no scope span',
+        () async {
+      fixture.options.enableLogs = true;
+
+      final client = fixture.getSut();
+      final log = givenLog();
+      final scope = Scope(fixture.options);
+
+      await client.captureLog(log, scope: scope);
+
+      final envelopePayloadJson = (fixture.transport).logs.first;
+      final logJson = envelopePayloadJson['items'].first;
+
+      expect(logJson['trace_id'], scope.propagationContext.traceId.toString());
+    });
+
+    test('$BeforeSendLogCallback returning null drops the log', () async {
+      fixture.options.enableLogs = true;
+      fixture.options.beforeSendLog = (log) => null;
+
+      final client = fixture.getSut();
+      final log = givenLog();
+
+      await client.captureLog(log);
+
+      expect((fixture.transport).logs, isEmpty);
+    });
+
+    test('$BeforeSendLogCallback returning a log modifies it', () async {
+      fixture.options.enableLogs = true;
+      fixture.options.beforeSendLog = (log) {
+        log.body = 'modified';
+        return log;
+      };
+
+      final client = fixture.getSut();
+      final log = givenLog();
+
+      await client.captureLog(log);
+
+      final envelopePayloadJson = (fixture.transport).logs.first;
+      final logJson = envelopePayloadJson['items'].first;
+
+      expect(logJson['body'], 'modified');
+    });
+
+    test('$BeforeSendLogCallback throwing is caught', () async {
+      fixture.options.enableLogs = true;
+      fixture.options.automatedTestMode = false;
+
+      fixture.options.beforeSendLog = (log) {
+        throw Exception('test');
+      };
+
+      final client = fixture.getSut();
+      final log = givenLog();
+
+      await client.captureLog(log);
+
+      final envelopePayloadJson = (fixture.transport).logs.first;
+      final logJson = envelopePayloadJson['items'].first;
+
+      expect(logJson['body'], 'test');
     });
   });